Second Commit

+++ src/main/java/RewardValue.java
@@ -1,24 +1,29 @@
 public class RewardValue {
 
     private double cashValue;
-    private int milesValue;
+    private double milesValue;
+
     private static final double MILES_TO_CASH_RATE = 0.0035;
 
+    // Constructor that accepts a cash value
     public RewardValue(double cashValue) {
         this.cashValue = cashValue;
-        this.milesValue = (int) ((int)cashValue / MILES_TO_CASH_RATE);
+        this.milesValue = cashValue / MILES_TO_CASH_RATE;
     }
 
+    // Constructor that accepts a value in miles
     public RewardValue(int milesValue) {
         this.milesValue = milesValue;
-        this.cashValue = milesValue*MILES_TO_CASH_RATE;
+        this.cashValue = milesValue * MILES_TO_CASH_RATE;
     }
 
+    // Method to get the cash value
     public double getCashValue() {
         return cashValue;
     }
 
-    public int getMilesValue() {
+    // Method to get the miles value
+    public double getMilesValue() {
         return milesValue;
     }
 }
